Implement logic
i.e. Now draw the rest of the owl.

Lot's happening here. The most important idea is folding over a
`{ rows : Int, columns : Int }`, which will go over the coordinates and
apply an accumulating function at all of them.

This is better than `List.range _ _` because no index list is actually
created. This saves memory and thus time.

Second thing is `buildDict`, which looks a little weird. I did it this
way, rather than actually returning the dict, to be able to  omit the
`Maybe.withDefault` later.

This computes the colors only once for every 2-char pair, previously it
would recompute it quite often.

Then I switched to BMP, and generate a 2D list. Both are for speed: bmp
does not do compression, which makes it much faster than png while still
producing good results. bmp uses a 2d list internally, so giving that as
an input is the fastest.